添加普攻跳字

+++ Assets/Scripts/Data/SkillConfig.cs
@@ -3,9 +3,9 @@ using System;
 [Serializable]
 public class SkillConfig
 {
-    public string Id;         
-    public string Name;      
-    public int CastTime;         // 前摇帧数
-    public string AnimName;   
-    public string SoundName; 
+    public string Id;
+    public string Name;
+    public float CastTime;
+    public string AnimName;
+    public string SoundName;
 }
